system auth module initial commit
