+++ app/openmrsid-common.js
@@ -9,17 +9,17 @@ CORE COMPONENTS
 */
 
 // load the modules (one at a time)
-exports.conf = require('./conf');
-exports.logger = require('./logger');
-exports.db = require('./db');
-exports.app = require('./app');
-exports.verification = require('./email-verification');
-exports.ldap = require('./ldap');
-exports.userNav = require('./user-nav'),
-exports.mid = require('./express-middleware'),
-exports.renderHelpers = require('./render-helpers'),
-exports.validate = require('./validate'),
-exports.environment = require('./environment');
+exports.conf          = require('./conf');
+exports.logger        = require('./logger');
+exports.db            = require('./db');
+exports.app           = require('./app');
+exports.verification  = require('./email-verification');
+exports.ldap          = require('./ldap');
+exports.userNav       = require('./user-nav');
+exports.mid           = require('./express-middleware');
+exports.renderHelpers = require('./render-helpers');
+exports.validate      = require('./validate');
+exports.environment   = require('./environment');
 
 /*
 MODULES
@@ -27,4 +27,4 @@ MODULES
 */
 
 // modules should define themselves at Common.module
-exports.module = {}
+exports.module = {};
