+++ app/system-modules/auth/index.js
@@ -1,0 +1,1 @@
+require('./lib');
