+++ app/system-modules/auth/lib/index.js
@@ -1,0 +1,1 @@
+require('./routes');
