+++ app/system-modules/auth/lib/routes/index.js
@@ -1,0 +1,1 @@
+require('./login');
